[Silver III] Title: 주유소, Time: 24 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -16,18 +16,18 @@ int main()
     cin >> N;
     long long* pDistance = (long long*)malloc(sizeof(long long) * (N - 1));
     long long* pGas = (long long*)malloc(sizeof(long long) * N);
-    for (int i = 0; i < N - 1; i++) {
+    for (long long i = 0; i < N - 1; i++) {
         cin >> pDistance[i];
         total += pDistance[i]; // 총 거리의 합
     }
-    for (int i = 0; i < N; i++) {
+    for (long long i = 0; i < N; i++) {
         cin >> pGas[i];
         if (pGas[i] < minOil) {
             minOil = pGas[i]; // 가장 저렴한 기름
         }
     }
 
-    for (int i = 0; i < N - 1; i++) {
+    for (long long i = 0; i < N - 1; i++) {
         if (pGas[i] == minOil) {
             money += total * minOil;
             break;
